move grid headers into grid subfolder
