+++ CMakeLists.txt
@@ -86,16 +86,16 @@ set(FASTSCAPELIB_HEADERS
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/flow_router_factory.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/flow_router.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/flow_routing.hpp
-  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/grid.hpp
+  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/grid/base.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/hillslope.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/iterators.hpp
-  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/profile_grid.hpp
-  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/raster_grid.hpp
+  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/grid/profile_grid.hpp
+  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/grid/raster_grid.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/sink_resolver_factory.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/sink_resolver.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/sinks.hpp
-  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/structured_grid.hpp
-  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/unstructured_mesh.hpp
+  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/grid/structured_grid.hpp
+  ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/grid/unstructured_mesh.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/utils.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/version.hpp
   ${FASTSCAPELIB_INCLUDE_DIR}/fastscapelib/xtensor_utils.hpp
