+++ benchmark/benchmark_flow_routing.cpp
@@ -1,8 +1,8 @@
 #include "benchmark_setup.hpp"
 
 #include "fastscapelib/consts.hpp"
-#include "fastscapelib/profile_grid.hpp"
-#include "fastscapelib/raster_grid.hpp"
+#include "fastscapelib/grid/profile_grid.hpp"
+#include "fastscapelib/grid/raster_grid.hpp"
 #include "fastscapelib/flow_routing.hpp"
 
 #include "xtensor/xfixed.hpp"
