+++ benchmark/benchmark_profile_grid.cpp
@@ -1,7 +1,7 @@
 #include "benchmark_setup.hpp"
 
 #include "fastscapelib/consts.hpp"
-#include "fastscapelib/profile_grid.hpp"
+#include "fastscapelib/grid/profile_grid.hpp"
 
 #include "xtensor/xfixed.hpp"
 #include "xtensor/xtensor.hpp"
