+++ include/fastscapelib/fastscapelib.hpp
@@ -13,11 +13,11 @@
 #include "fastscapelib/flow_graph.hpp"
 #include "fastscapelib/flow_router.hpp"
 #include "fastscapelib/sink_resolver.hpp"
-#include "fastscapelib/grid.hpp"
-#include "fastscapelib/unstructured_mesh.hpp"
-#include "fastscapelib/structured_grid.hpp"
-#include "fastscapelib/profile_grid.hpp"
-#include "fastscapelib/raster_grid.hpp"
+#include "fastscapelib/grid/base.hpp"
+#include "fastscapelib/grid/unstructured_mesh.hpp"
+#include "fastscapelib/grid/structured_grid.hpp"
+#include "fastscapelib/grid/profile_grid.hpp"
+#include "fastscapelib/grid/raster_grid.hpp"
 #include "fastscapelib/hillslope.hpp"
 #include "fastscapelib/sinks.hpp"
