+++ test/test_flow_routing.cpp
@@ -1,5 +1,5 @@
 #include "fastscapelib/utils.hpp"
-#include "fastscapelib/raster_grid.hpp"
+#include "fastscapelib/grid/raster_grid.hpp"
 #include "fastscapelib/flow_routing.hpp"
 
 #include "xtensor/xtensor.hpp"
